Changed Resources folders structure
