+++ Microsoft.Recognizers.Resources/BaseDateTime.tt
@@ -1,0 +1,7 @@
+﻿<#@ template debug="true" hostspecific="true" language="C#" #>
+<#
+	this.DataFilename = @"Patterns\Base-DateTime.yaml";
+	this.Language = null;
+	this.ClassName = "BaseDateTime";
+#>
+<#@ include file=".\CommonDefinitions.ttinclude"#>
