+++ Microsoft.Recognizers.Resources/BaseNumbers.tt
@@ -1,0 +1,7 @@
+﻿<#@ template debug="true" hostspecific="true" language="C#" #>
+<#
+	this.DataFilename = @"Patterns\Base-Numbers.yaml";
+	this.Language = null;
+	this.ClassName = "BaseNumbers";
+#>
+<#@ include file=".\CommonDefinitions.ttinclude"#>
