--- Microsoft.Recognizers.Resources/English/DateTimeDefinition.tt
@@ -1,7 +1,0 @@
-﻿﻿<#@ template debug="true" hostspecific="true" language="C#" #>
-<#
-	this.DataFilename = @"Common\DateTime-English.yaml";
-	this.Language = "English";
-	this.ClassName = "DateTimeDefinition";
-#>
-<#@ include file="..\NumericBase.ttinclude"#>
