+++ Microsoft.Recognizers.Resources/English/NumbersDefinitions.tt
@@ -1,0 +1,7 @@
+﻿<#@ template debug="true" hostspecific="true" language="C#" #>
+<#
+	this.DataFilename = @"Patterns\English\English-Numbers.yaml";
+	this.Language = "English";
+	this.ClassName = "NumbersDefinitions";
+#>
+<#@ include file="..\CommonDefinitions.ttinclude"#>
