+++ Microsoft.Recognizers.Resources/English/NumbersWithUnitDefinitions.tt
@@ -1,0 +1,7 @@
+﻿<#@ template debug="true" hostspecific="true" language="C#" #>
+<#
+	this.DataFilename = @"Patterns\English\English-NumbersWithUnit.yaml";
+	this.Language = "English";
+	this.ClassName = "NumericWithUnit";
+#>
+<#@ include file="..\CommonDefinitions.ttinclude"#>
