--- Microsoft.Recognizers.Resources/English/Numeric.tt
@@ -1,7 +1,0 @@
-﻿<#@ template debug="true" hostspecific="true" language="C#" #>
-<#
-	this.DataFilename = @"Common\Numeric-English.yaml";
-	this.Language = "English";
-	this.ClassName = "Numeric";
-#>
-<#@ include file="..\NumericBase.ttinclude"#>
