--- Microsoft.Recognizers.Resources/English/NumericWithUnit.tt
@@ -1,7 +1,0 @@
-﻿<#@ template debug="true" hostspecific="true" language="C#" #>
-<#
-	this.DataFilename = @"Common\NumericWithUnit-English.yaml";
-	this.Language = "English";
-	this.ClassName = "NumericWithUnit";
-#>
-<#@ include file="..\NumericBase.ttinclude"#>
