+++ Patterns/English/English-DateTime.yaml
@@ -152,8 +152,8 @@ PmRegex: !simpleRegex
 AmRegex: !simpleRegex
   def: (?<am>(in the\s+)?morning)
 LessThanOneHour: !nestedRegex
-  def: (?<lth>(a\s+)?quarter|three quarter(s)?|half( an hour)?|{CommonDateTime.DeltaMinuteRegex}(\s+(minute|minutes|min|mins))|{DeltaMinuteNumRegex}(\s+(minute|minutes|min|mins)))
-  references: [ CommonDateTime.DeltaMinuteRegex, DeltaMinuteNumRegex ]
+  def: (?<lth>(a\s+)?quarter|three quarter(s)?|half( an hour)?|{BaseDateTime.DeltaMinuteRegex}(\s+(minute|minutes|min|mins))|{DeltaMinuteNumRegex}(\s+(minute|minutes|min|mins)))
+  references: [ BaseDateTime.DeltaMinuteRegex, DeltaMinuteNumRegex ]
 EngTimeRegex: !nestedRegex
   def: (?<engtime>{HourNumRegex}\s+({MinuteNumRegex}|(?<tens>twenty|thirty|forty|fourty|fifty)\s+{MinuteNumRegex}))
   references: [ HourNumRegex, MinuteNumRegex ]
@@ -164,28 +164,28 @@ TimeSuffix: !nestedRegex
   def: (?<suffix>{AmRegex}|{PmRegex}|{OclockRegex})
   references: [ AmRegex, PmRegex, OclockRegex ]
 BasicTime: !nestedRegex
-  def: (?<basictime>{EngTimeRegex}|{HourNumRegex}|{CommonDateTime.HourRegex}:{CommonDateTime.MinuteRegex}(:{CommonDateTime.SecondRegex})?|{CommonDateTime.HourRegex})
-  references: [ EngTimeRegex, HourNumRegex, CommonDateTime.HourRegex, CommonDateTime.MinuteRegex, CommonDateTime.SecondRegex ]
+  def: (?<basictime>{EngTimeRegex}|{HourNumRegex}|{BaseDateTime.HourRegex}:{BaseDateTime.MinuteRegex}(:{BaseDateTime.SecondRegex})?|{BaseDateTime.HourRegex})
+  references: [ EngTimeRegex, HourNumRegex, BaseDateTime.HourRegex, BaseDateTime.MinuteRegex, BaseDateTime.SecondRegex ]
 AtRegex: !nestedRegex
-  def: \b(?<=\bat\s+)({EngTimeRegex}|{HourNumRegex}|{CommonDateTime.HourRegex})\b
-  references: [ EngTimeRegex, HourNumRegex, CommonDateTime.HourRegex ]
+  def: \b(?<=\bat\s+)({EngTimeRegex}|{HourNumRegex}|{BaseDateTime.HourRegex})\b
+  references: [ EngTimeRegex, HourNumRegex, BaseDateTime.HourRegex ]
 IshRegex: !nestedRegex
-  def: '{CommonDateTime.HourRegex}(-|——)?ish|noonish|noon'
-  references: [ CommonDateTime.HourRegex ]
+  def: '{BaseDateTime.HourRegex}(-|——)?ish|noonish|noon'
+  references: [ BaseDateTime.HourRegex ]
 TimeUnitRegex: !simpleRegex
   def: (?<unit>hours|hour|hrs|hr|h|minutes|minute|mins|min|seconds|second|secs|sec)\b
 ConnectNumRegex: !nestedRegex
-  def: '{CommonDateTime.HourRegex}(?<min>00|01|02|03|04|05|06|07|08|09|10|11|12|13|14|15|16|17|18|19|20|21|22|23|24|25|26|27|28|29|30|31|32|33|34|35|36|37|38|39|40|41|42|43|44|45|46|47|48|49|50|51|52|53|54|55|56|57|58|59)\s*{DescRegex}'
-  references: [ CommonDateTime.HourRegex, DescRegex ]
+  def: '{BaseDateTime.HourRegex}(?<min>00|01|02|03|04|05|06|07|08|09|10|11|12|13|14|15|16|17|18|19|20|21|22|23|24|25|26|27|28|29|30|31|32|33|34|35|36|37|38|39|40|41|42|43|44|45|46|47|48|49|50|51|52|53|54|55|56|57|58|59)\s*{DescRegex}'
+  references: [ BaseDateTime.HourRegex, DescRegex ]
 TimeRegex1: !nestedRegex
-  def: (\b{TimePrefix}\s+)?({EngTimeRegex}|{HourNumRegex}|{CommonDateTime.HourRegex})\s*{DescRegex}
-  references: [ TimePrefix, EngTimeRegex, HourNumRegex, CommonDateTime.HourRegex, DescRegex ]
+  def: (\b{TimePrefix}\s+)?({EngTimeRegex}|{HourNumRegex}|{BaseDateTime.HourRegex})\s*{DescRegex}
+  references: [ TimePrefix, EngTimeRegex, HourNumRegex, BaseDateTime.HourRegex, DescRegex ]
 TimeRegex2: !nestedRegex
-  def: (\b{TimePrefix}\s+)?(T)?{CommonDateTime.HourRegex}(\s*)?:(\s*)?{CommonDateTime.MinuteRegex}((\s*)?:(\s*)?{CommonDateTime.SecondRegex})?((\s*{DescRegex})|\b)
-  references: [ TimePrefix, CommonDateTime.HourRegex, CommonDateTime.MinuteRegex, CommonDateTime.SecondRegex, DescRegex ]
+  def: (\b{TimePrefix}\s+)?(T)?{BaseDateTime.HourRegex}(\s*)?:(\s*)?{BaseDateTime.MinuteRegex}((\s*)?:(\s*)?{BaseDateTime.SecondRegex})?((\s*{DescRegex})|\b)
+  references: [ TimePrefix, BaseDateTime.HourRegex, BaseDateTime.MinuteRegex, BaseDateTime.SecondRegex, DescRegex ]
 TimeRegex3: !nestedRegex
-  def: (\b{TimePrefix}\s+)?{CommonDateTime.HourRegex}\.{CommonDateTime.MinuteRegex}(\s*{DescRegex})
-  references: [ TimePrefix, CommonDateTime.HourRegex, CommonDateTime.MinuteRegex, DescRegex ]
+  def: (\b{TimePrefix}\s+)?{BaseDateTime.HourRegex}\.{BaseDateTime.MinuteRegex}(\s*{DescRegex})
+  references: [ TimePrefix, BaseDateTime.HourRegex, BaseDateTime.MinuteRegex, DescRegex ]
 TimeRegex4: !nestedRegex
   def: \b{TimePrefix}\s+{BasicTime}(\s*{DescRegex})?\s+{TimeSuffix}\b
   references: [ TimePrefix, BasicTime, DescRegex, TimeSuffix ]
@@ -246,11 +246,11 @@ TimeOfTodayBeforeRegex: !nestedRegex
   def: '{DateTimeSpecificNightRegex}(\s*,)?(\s+(at|for))?\s*$'
   references: [ DateTimeSpecificNightRegex ]
 SimpleTimeOfTodayAfterRegex: !nestedRegex
-  def: ({HourNumRegex}|{CommonDateTime.HourRegex})\s*(,\s*)?(in\s+)?{DateTimeSpecificNightRegex}
-  references: [ HourNumRegex, CommonDateTime.HourRegex, DateTimeSpecificNightRegex ]
+  def: ({HourNumRegex}|{BaseDateTime.HourRegex})\s*(,\s*)?(in\s+)?{DateTimeSpecificNightRegex}
+  references: [ HourNumRegex, BaseDateTime.HourRegex, DateTimeSpecificNightRegex ]
 SimpleTimeOfTodayBeforeRegex: !nestedRegex
-  def: '{DateTimeSpecificNightRegex}(\s*,)?(\s+(at|for))?\s*({HourNumRegex}|{CommonDateTime.HourRegex})'
-  references: [ DateTimeSpecificNightRegex, HourNumRegex, CommonDateTime.HourRegex ]
+  def: '{DateTimeSpecificNightRegex}(\s*,)?(\s+(at|for))?\s*({HourNumRegex}|{BaseDateTime.HourRegex})'
+  references: [ DateTimeSpecificNightRegex, HourNumRegex, BaseDateTime.HourRegex ]
 TheEndOfRegex: !simpleRegex
   def: (the\s+)?end of(\s+the)?\s*$
 PeriodNightRegex: !simpleRegex
